+++ examples/xeyes.rs
@@ -4,10 +4,10 @@ extern crate x11rb;
 
 use x11rb::connection::{Connection, RequestConnection as _};
 use x11rb::errors::{ConnectionError, ReplyOrIdError};
-use x11rb::generated::shape::{self, ConnectionExt as _};
-use x11rb::generated::xproto::*;
+use x11rb::shape::{self, ConnectionExt as _};
 use x11rb::wrapper::ConnectionExt as _;
 use x11rb::x11_utils::Event;
+use x11rb::xproto::*;
 use x11rb::COPY_DEPTH_FROM_PARENT;
 
 const PUPIL_SIZE: i16 = 50;