+++ src/x11_utils.rs
@@ -41,7 +41,7 @@ pub trait Event {
     ///
     /// Not all packets contain a sequence number, so this function returns an `Option`.
     fn raw_sequence_number(&self) -> Option<u16> {
-        use crate::generated::xproto::KEYMAP_NOTIFY_EVENT;
+        use crate::xproto::KEYMAP_NOTIFY_EVENT;
         match self.response_type() {
             KEYMAP_NOTIFY_EVENT => None,
             _ => {
@@ -61,7 +61,7 @@ pub struct GenericEvent<B: AsRef<[u8]>>(B);
 
 impl<B: AsRef<[u8]>> GenericEvent<B> {
     pub fn new(value: B) -> Result<Self, ParseError> {
-        use super::generated::xproto::GE_GENERIC_EVENT;
+        use super::xproto::GE_GENERIC_EVENT;
         let value_slice = value.as_ref();
         if value_slice.len() < 32 {
             return Err(ParseError::ParseError);
@@ -330,7 +330,7 @@ macro_rules! bitmask_binop {
 ///
 /// If we need to use multiple atoms, one would normally write code such as
 /// ```
-/// # use x11rb::generated::xproto::{ATOM, ConnectionExt, InternAtomReply};
+/// # use x11rb::xproto::{ATOM, ConnectionExt, InternAtomReply};
 /// # use x11rb::errors::{ConnectionError, ReplyError};
 /// # use x11rb::cookie::Cookie;
 /// #[allow(non_snake_case)]
@@ -397,9 +397,9 @@ macro_rules! atom_manager {
         // Cookie version
         #[allow(non_snake_case)]
         #[derive(Debug)]
-        $vis struct $cookie_name<'a, C: $crate::generated::xproto::ConnectionExt> {
+        $vis struct $cookie_name<'a, C: $crate::xproto::ConnectionExt> {
             $(
-                $field_name: $crate::cookie::Cookie<'a, C, $crate::generated::xproto::InternAtomReply>,
+                $field_name: $crate::cookie::Cookie<'a, C, $crate::xproto::InternAtomReply>,
             )*
         }
 
@@ -408,12 +408,12 @@ macro_rules! atom_manager {
         #[derive(Debug, Clone, Copy)]
         $vis struct $struct_name {
             $(
-                $vis $field_name: $crate::generated::xproto::ATOM,
+                $vis $field_name: $crate::xproto::ATOM,
             )*
         }
 
         impl $struct_name {
-            $vis fn new<C: $crate::generated::xproto::ConnectionExt>(
+            $vis fn new<C: $crate::xproto::ConnectionExt>(
                 conn: &C,
             ) -> ::std::result::Result<$cookie_name<'_, C>, $crate::errors::ConnectionError> {
                 Ok($cookie_name {
@@ -427,7 +427,7 @@ macro_rules! atom_manager {
             }
         }
 
-        impl<'a, C: $crate::generated::xproto::ConnectionExt> $cookie_name<'a, C> {
+        impl<'a, C: $crate::xproto::ConnectionExt> $cookie_name<'a, C> {
             $vis fn reply(self) -> ::std::result::Result<$struct_name, $crate::errors::ReplyError<C::Buf>> {
                 Ok($struct_name {
                     $(