+++ tests/regression_tests.rs
@@ -6,12 +6,12 @@ use std::ops::Deref;
 use x11rb::connection::{BufWithFds, DiscardMode, RequestConnection, RequestKind, SequenceNumber};
 use x11rb::cookie::{Cookie, CookieWithFds, VoidCookie};
 use x11rb::errors::{ConnectionError, ParseError, ReplyError};
-use x11rb::generated::xproto::{
+use x11rb::utils::RawFdContainer;
+use x11rb::x11_utils::{GenericError, Serialize, TryParse};
+use x11rb::xproto::{
     ClientMessageData, ConnectionExt, KeymapNotifyEvent, QueryExtensionReply, Segment,
     SetupAuthenticate,
 };
-use x11rb::utils::RawFdContainer;
-use x11rb::x11_utils::{GenericError, Serialize, TryParse};
 
 #[derive(Debug)]
 struct SavedRequest {
@@ -169,7 +169,7 @@ fn test_poly_segment() -> Result<(), ReplyError<Vec<u8>>> {
     conn.poly_segment(drawable, gc, &segments)?;
 
     let mut expected = Vec::new();
-    expected.push(x11rb::generated::xproto::POLY_SEGMENT_REQUEST);
+    expected.push(x11rb::xproto::POLY_SEGMENT_REQUEST);
     expected.push(0); // padding
     expected.extend(&length.to_ne_bytes()); // length, not in the xml
     expected.extend(&drawable.to_ne_bytes());
@@ -195,9 +195,10 @@ fn test_big_requests() -> Result<(), ConnectionError> {
     conn.poly_text16(drawable, gc, x, y, &big_buffer)?;
 
     let mut expected = Vec::new();
-    expected.push(x11rb::generated::xproto::POLY_TEXT16_REQUEST);
-    expected.push(0); // padding
-                      // Length of zero: we use big requests
+    expected.push(x11rb::xproto::POLY_TEXT16_REQUEST);
+    // padding
+    expected.push(0);
+    // Length of zero: we use big requests
     expected.push(0);
     expected.push(0);
     // Actual length
@@ -247,7 +248,7 @@ fn test_send_event() -> Result<(), ConnectionError> {
     conn.send_event(propagate, destination, event_mask, event)?;
 
     let mut expected = Vec::new();
-    expected.push(x11rb::generated::xproto::SEND_EVENT_REQUEST);
+    expected.push(x11rb::xproto::SEND_EVENT_REQUEST);
     expected.push(propagate as _);
     expected.extend(&((12u16 + 32u16) / 4).to_ne_bytes());
     expected.extend(&destination.to_ne_bytes());